requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
google-generativeai==0.7.0
